2, 3단계 - 로그인 리팩터링, 관리자 기능 보호 (#17)
* docs: 기능 요구사항 업데이트

* feat: 로그인 기능 구현

* feat: 내 정보 조회하기 기능 구현

* feat: 비로그인 사용자는 예약이 불가능하다

* feat: 자신의 예약이 아닌 경우 예약 취소가 불가능하다.

* feat: 관리자 역할을 추가한다

* feat: 관리자 기능을 보호한다

* feat: 관리자를 등록하도록 하기 보다는 애플리케이션이 동작할 때 관리자는 추가될 수 있도록 한다

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -8,6 +8,7 @@ import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class ReservationService {
@@ -21,7 +22,7 @@ public class ReservationService {
         this.scheduleDao = scheduleDao;
     }
 
-    public Long create(ReservationRequest reservationRequest) {
+    public Long create(ReservationRequest reservationRequest, Long memberId) {
         Schedule schedule = scheduleDao.findById(reservationRequest.getScheduleId());
         if (schedule == null) {
             throw new NullPointerException();
@@ -34,7 +35,8 @@ public class ReservationService {
 
         Reservation newReservation = new Reservation(
                 schedule,
-                reservationRequest.getName()
+                reservationRequest.getName(),
+                memberId
         );
 
         return reservationDao.save(newReservation);
@@ -49,10 +51,14 @@ public class ReservationService {
         return reservationDao.findAllByThemeIdAndDate(themeId, date);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, Long memberId) {
         Reservation reservation = reservationDao.findById(id);
         if (reservation == null) {
             throw new NullPointerException();
+        }
+
+        if (!Objects.equals(reservation.getMemberId(), memberId)) {
+            throw new IllegalArgumentException("자신의 예약이 아닌 경우 예약 취소가 불가능합니다.");
         }
 
         reservationDao.deleteById(id);
